refactor: change the query package name to queries
