+++ .golangci.yaml
@@ -378,11 +378,11 @@ run:
   build-tags:
     - mytag
   # If set, we pass it to "go list -mod={option}". From "go help modules":
-  # If invoked with -mod=readonly, the go query is disallowed from the implicit
+  # If invoked with -mod=readonly, the go queries is disallowed from the implicit
   # automatic updating of go.mod described above. Instead, it fails when any changes
   # to go.mod are needed. This setting is most useful to check that go.mod does
   # not need updates, such as in a continuous integration and testing system.
-  # If invoked with -mod=vendor, the go query assumes that the vendor
+  # If invoked with -mod=vendor, the go queries assumes that the vendor
   # directory holds the correct copies of dependencies and ignores
   # the dependency descriptions in go.mod.
   #
