+++ internal/modules/user/features/login/v1/application/queries/login_query.go
@@ -1,4 +1,4 @@
-package query
+package queries
 
 type LoginQueryV1 struct {
 	Email    string
