+++ internal/modules/user/features/me/v1/application/queries/get_me_handler.go
@@ -1,4 +1,4 @@
-package query
+package queries
 
 import (
 	"context"
