+++ internal/modules/user/features/refreshToken/v1/application/queries/refresh_token_handler.go
@@ -1,4 +1,4 @@
-package query
+package queries
 
 import (
 	"context"
