+++ internal/modules/user/features/refreshToken/v1/application/queries/refresh_token_query.go
@@ -1,4 +1,4 @@
-package query
+package queries
 
 type RefreshTokenQueryV1 struct {
 	RefreshToken string
