+++ internal/modules/user/module.go
@@ -6,8 +6,8 @@ import (
 	userDomainPorts "github.com/racibaz/go-arch/internal/modules/user/domain/ports"
 	loginV1Query "github.com/racibaz/go-arch/internal/modules/user/features/login/v1/application/queries"
 	logoutCommands "github.com/racibaz/go-arch/internal/modules/user/features/logout/v1/application/commands"
-	meQueryV1 "github.com/racibaz/go-arch/internal/modules/user/features/me/v1/application/query"
-	refreshTokenQueryV1 "github.com/racibaz/go-arch/internal/modules/user/features/refreshToken/v1/application/query"
+	meQueryV1 "github.com/racibaz/go-arch/internal/modules/user/features/me/v1/application/queries"
+	refreshTokenQueryV1 "github.com/racibaz/go-arch/internal/modules/user/features/refreshToken/v1/application/queries"
 	"github.com/racibaz/go-arch/internal/modules/user/features/signup/v1/application/commands"
 	"github.com/racibaz/go-arch/pkg/ddd"
 	"github.com/racibaz/go-arch/pkg/logger"
