+++ internal/modules/user/module_test.go
@@ -6,8 +6,8 @@ import (
 	sharedPortsMocks "github.com/racibaz/go-arch/internal/modules/shared/testing/mocks/application/ports"
 	userQueries "github.com/racibaz/go-arch/internal/modules/user/features/login/v1/application/queries"
 	logoutCommands "github.com/racibaz/go-arch/internal/modules/user/features/logout/v1/application/commands"
-	meQueries "github.com/racibaz/go-arch/internal/modules/user/features/me/v1/application/query"
-	refreshTokenQueries "github.com/racibaz/go-arch/internal/modules/user/features/refreshToken/v1/application/query"
+	meQueries "github.com/racibaz/go-arch/internal/modules/user/features/me/v1/application/queries"
+	refreshTokenQueries "github.com/racibaz/go-arch/internal/modules/user/features/refreshToken/v1/application/queries"
 	userCommands "github.com/racibaz/go-arch/internal/modules/user/features/signup/v1/application/commands"
 	userPortsMocks "github.com/racibaz/go-arch/internal/modules/user/testing/mocks/domain/ports"
 	"github.com/racibaz/go-arch/pkg/ddd"
