+++ internal/modules/user/routes.go
@@ -10,9 +10,9 @@ import (
 	logoutV1Endpoint "github.com/racibaz/go-arch/internal/modules/user/features/logout/v1/adapters/endpoints"
 	logoutV1Command "github.com/racibaz/go-arch/internal/modules/user/features/logout/v1/application/commands"
 	meV1Endpoint "github.com/racibaz/go-arch/internal/modules/user/features/me/v1/adapters/endpoints"
-	meV1Query "github.com/racibaz/go-arch/internal/modules/user/features/me/v1/application/query"
+	meV1Query "github.com/racibaz/go-arch/internal/modules/user/features/me/v1/application/queries"
 	refreshTokenV1Endpoint "github.com/racibaz/go-arch/internal/modules/user/features/refreshToken/v1/adapters/endpoints"
-	refreshTokenV1Query "github.com/racibaz/go-arch/internal/modules/user/features/refreshToken/v1/application/query"
+	refreshTokenV1Query "github.com/racibaz/go-arch/internal/modules/user/features/refreshToken/v1/application/queries"
 	registeringUserV1Endpoint "github.com/racibaz/go-arch/internal/modules/user/features/signup/v1/adapters/endpoints"
 	commandsV1Endpoint "github.com/racibaz/go-arch/internal/modules/user/features/signup/v1/application/commands"
 	"github.com/racibaz/go-arch/internal/modules/user/infrastructure/hashing"
